+++ test/test_util.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 import platform
 import pytest
 import requests
-from ghreq import get_github_api_url, make_user_agent
+from ghreq import get_github_api_url, get_github_graphql_url, make_user_agent
 
 
 def test_get_github_api_url_no_envvar(monkeypatch: pytest.MonkeyPatch) -> None:
@@ -18,6 +18,21 @@ def test_get_github_api_url_empty_envvar(monkeypatch: pytest.MonkeyPatch) -> Non
 def test_get_github_api_url_from_envvar(monkeypatch: pytest.MonkeyPatch) -> None:
     monkeypatch.setenv("GITHUB_API_URL", "https://github.example.com/api")
     assert get_github_api_url() == "https://github.example.com/api"
+
+
+def test_get_github_graphql_url_no_envvar(monkeypatch: pytest.MonkeyPatch) -> None:
+    monkeypatch.delenv("GITHUB_GRAPHQL_URL", raising=False)
+    assert get_github_graphql_url() == "https://api.github.com/graphql"
+
+
+def test_get_github_graphql_url_empty_envvar(monkeypatch: pytest.MonkeyPatch) -> None:
+    monkeypatch.setenv("GITHUB_GRAPHQL_URL", "")
+    assert get_github_graphql_url() == "https://api.github.com/graphql"
+
+
+def test_get_github_graphql_url_from_envvar(monkeypatch: pytest.MonkeyPatch) -> None:
+    monkeypatch.setenv("GITHUB_GRAPHQL_URL", "https://github.example.com/api/graphql")
+    assert get_github_graphql_url() == "https://github.example.com/api/graphql"
 
 
 @pytest.mark.parametrize(